pyproject.toml
[tool.poetry]
name = "niftivis"
version = "2020.06.1"
description = "Convert NIFTI to PNG in various ways."
authors = ["Jon Stutters <j.stutters@ucl.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
nibabel = "^3.0.0"
pillow = "^6.2.1"
scikit-image = "^0.16.2"
click = "^7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pylint = "^2.5.3"

[tool.poetry.scripts]
niftivis = "niftivis.niftivis:make_thumbnails"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
